Socket class modifications.
1. Move socket files to `net/socket.*` so that we can use sockets
   elsewhere in the code, not only in HTTPD.
2. Add methods to socket class for byte data manipulation.
3. Add additional testing to socket methods. (e.g. Do not call the
   wrapped functions on closed or non-initialized sockets.)
