+++ CMakeLists.txt
@@ -87,6 +87,7 @@ SET(
     src/http/cookie.cc
     src/http/method.cc
     src/http/version.cc
+    src/net/socket.cc
     src/sapi/request.cc
     src/sapi/response.cc
 )
@@ -123,7 +124,6 @@ IF (TEMPEARLY_ENABLE_HTTPD_SAPI)
         src/sapi/httpd/httpd-response.cc
         src/sapi/httpd/main.cc
         src/sapi/httpd/server.cc
-        src/sapi/httpd/socket.cc
         ${INTERPRETER_SOURCES}
     )
 ENDIF()
