+++ src/defines.h
@@ -64,6 +64,7 @@ namespace tempearly
     class Result;
     class Scope;
     class Script;
+    class Socket;
     class String;
     class StringBuilder;
     class Token;
